migrate DriveCommands into Drive subsystem
